+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New features
 Fixes
 
 - Skip rendering if app state is not changed by an action
+- `o "div.main", class: 'hovered'` should yield `<div class='main hovered'>`
 
 ## v0.1.0 2018-06-01